requirements.txt
# Core Neo4j dependencies
neo4j>=5.13.0
neo4j-driver>=5.13.0

# Data processing and validation
pandas>=2.0.0
pydantic>=2.0.0
numpy>=1.24.0

# Configuration and environment
PyYAML>=6.0.0
python-dotenv>=1.0.0

# CLI and utilities
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Sample data generation
faker>=20.0.0

# Optional: For advanced data processing
# scipy>=1.11.0
# scikit-learn>=1.3.0

# Optional: For API development
# fastapi>=0.103.0
# uvicorn>=0.23.0

# Optional: For container orchestration
# kubernetes>=27.2.0
